Update to ember:try versions.
Due to the version of ember-cli and thus ember:try using version
compatibility only would go up to the 2.18 release line and then run
beta/canary version for the 3.0 line.

Now that ember 3 is a release candidate we want to actually run the
tests against that release version too.

So for now until the project is upgraded to a newer version of ember
going to just hard code some versions.

Using the ember docs went with the last few LTS releases (2.10,2.12,2.16 and 2.18) and then both
minor versions of the 3.x (3.0 and 3.1) train.